solve 2.11
