+++ 2.6-2.11/src/Components/Filter.js
@@ -1,0 +1,8 @@
+const Filter = ({ search, addSearch }) => {
+    return (
+      <div>
+        search: <input value={search} onChange={addSearch}></input>
+      </div>
+    )
+  }
+  export default Filter
