+++ 2.6-2.11/src/Components/Person.js
@@ -1,0 +1,8 @@
+const Person = ({ person }) => {
+    return (
+      <div>
+        <li>{person.name} {person.number}</li>
+      </div>
+    )
+  }
+  export default Person
