+++ 2.6-2.11/src/Components/Persons.js
@@ -1,0 +1,13 @@
+import Person from './Person'
+const Persons = ({ persons, search }) => {
+    let arr = persons.filter(person => person.name.toLowerCase().includes(search.toLowerCase()) || person.number.includes(search))
+  
+    return (
+      <div>
+        {arr.map(person =>
+          <Person key={person.name} person={person} />
+        )}
+      </div>
+    )
+  }
+export default Persons
